[dropshot-endpoint] move channel and endpoint code into their own files (#1006)
I'm going to add a `dropshot::server` macro next to these, and it's big enough
that it warrants its own file. The others should similarly live in their own
files as well.

I also moved some utility code shared between the endpoint and the server into
another file.

This is mostly big-hammer refactoring for now with no functional changes. The
server macro PR will do some more fine-grained refactoring on top of this.
